Re-exec gamescope to launch without preloading gameoverlayrenderer.so
Currently, gamescope removes gameoverlayrenderer.so from LD_PRELOAD
before launching the child process.  This change reverses the logic.  It
launches gamescope with a modified LD_PRELOAD that does not contain
gameoverlayrenderer.so, but does launch the child process with it.  To
do that, we have to re-exec gamescope with the modified LD_PRELOAD env
variable and store the original value of LD_PRELOAD in another
environment variable so we can pass it on to the child process.